Split FileInputRow into a container and functional component
