+++ src/components/__stories__/FileInputRow.stories.js
@@ -11,14 +11,16 @@ storiesOf('FileInputRow', module)
   .add('No file', () => (
     <FileInputRow
       filename={null}
-      onLoading={() => undefined}
-      onLoaded={(s: ?string, o: ?Object) => undefined}
+      dataPaths={null}
+      isDragging={false}
+      onStatsFilePicked={(path: string | null) => undefined}
     />
   ))
   .add('File loaded', () => (
     <FileInputRow
       filename={'stats.json'}
-      onLoading={() => undefined}
-      onLoaded={(s: ?string, o: ?Object) => undefined}
+      dataPaths={null}
+      isDragging={false}
+      onStatsFilePicked={(path: string | null) => undefined}
     />
   ));
